Add support for specifying a build directory
When a build directory is explicitly specified, it's not removed after
the build. This is the first step towards supporting incremental builds.

+++ tuxmake/build.py
@@ -56,6 +56,7 @@ class Build:
         self,
         tree=".",
         output_dir=None,
+        build_dir=None,
         target_arch=None,
         toolchain=None,
         wrapper=None,
@@ -76,8 +77,13 @@ class Build:
             self.output_dir = output_dir
             os.mkdir(self.output_dir)
 
-        self.build_dir = self.output_dir / "tmp"
-        os.mkdir(self.build_dir)
+        if build_dir:
+            self.build_dir = Path(build_dir)
+            self.keep_build_dir = True
+        else:
+            self.build_dir = self.output_dir / "tmp"
+            self.build_dir.mkdir()
+            self.keep_build_dir = False
 
         self.target_arch = target_arch and Architecture(target_arch) or host_arch
         self.toolchain = toolchain and Toolchain(toolchain) or NoExplicitToolchain()
@@ -313,7 +319,8 @@ class Build:
         self.logger.terminate()
 
     def cleanup(self):
-        shutil.rmtree(self.build_dir)
+        if not self.keep_build_dir:
+            shutil.rmtree(self.build_dir)
 
     def run(self):
         self.validate()
